# documents/swagger.yaml
openapi: 3.0.0
info:
  description: "This is API documentation for Vtiger custmer portal service. This is only client for Vtiger API. This is a place for your customers to see and update tickets, view past quotes and invoices, and share files. The customer portal is a self-service portal where customers can get their business-related questions answered."
  version: 1.0.2
  title: Vtiger Go Customer Portal
  termsOfService: http://swagger.io/terms/
  contact:
    email: se@sergeyem.ru
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/modules/{moduleName}":
    get:
      tags:
        - module
      summary: Module Info
      description: Get all information about module by it name
      operationId: describeModule
      security:
        - bearerAuth: []
      parameters:
        - name: moduleName
          in: path
          description: Name of module to return
          required: true
          schema:
            type: string
            format: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Module"
        "400":
          description: Invalid Name supplied
        "404":
          description: Module not found
        "500":
          description: Vtiger error response
...
  "/custom-modules/{module}":
    get:
      tags:
        - custom-modules
      summary: Get All Records
      description: Get all records from custom module
      operationId: getCustomModuleRecords
      security:
        - bearerAuth: []
      parameters:
        - name: module
          in: path
          description: Module name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                  count:
                    type: number
                    example: 100
                  page:
                    type: number
                    example: 1
                  size:
                    type: number
                    example: 100
        "400":
          description: Invalid ID supplied
        "403":
          description: Operation not permitted
  "/custom-modules/{module}/{id}":
    get:
      tags:
        - custom-modules
      summary: Get Record
      description: Get custom module record by ID
      operationId: getCustomModuleRecordById
      security:
        - bearerAuth: []
      parameters:
        - name: module
          in: path
          description: Module name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Record ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid ID supplied
        "403":
          description: Operation not permitted
...
servers:
  - url: http://127.0.0.1:4045/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
...
  schemas:
...
    Module:
      type: object
      properties:
        label:
          type: string
          example: "Invoices"
        name:
          type: string
          example: "Invoice"
        createable:
          type: boolean
          example: true
        updateable:
          type: boolean
          example: true
        deleteable:
          type: boolean
          example: true
        retrieveable:
          type: boolean
          example: true
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ModuleField'
          example:
            - name: "subject"
              label: "Subject"
              mandatory: true
              isunique: true
              nullable: false
              editable: true
              default: ""
            - name: "duedate"
              label: "Due Date"
              mandatory: false
              isunique: false
              nullable: true
              editable: true
              default: "default2"
        idPrefix:
          type: string
          example: "2"
        isEntity:
          type: boolean
          example: true
        allowDuplicates:
          type: boolean
          example: false
        labelFields:
          type: string
          example: "subject"
    ModuleField:
      type: object
      properties:
        name:
          type: string
          example: "invoicestatus"
        label:
          type: string
          example: "Invoice Status"
        mandatory:
          type: boolean
          example: true
        isunique:
          type: boolean
          example: true
        nullable:
          type: boolean
          example: false
        editable:
          type: boolean
          example: true
        default:
          type: string
          example: "Created"
...